+++ src/ui/barrel.ts
@@ -1,3 +1,4 @@
 export * from './THEOplayerDefaultUi';
 export * from './THEOplayerTheme';
 export * from './components/barrel';
+export * from './hooks/barrel';