+++ src/ui/components/button/ChromecastButton.tsx
@@ -1,15 +1,13 @@
-import React, { PureComponent, type ReactNode } from 'react';
+import React, { type ReactNode } from 'react';
 import { CastButton } from 'react-native-google-cast';
-import type { CastEvent } from 'react-native-theoplayer';
-import { CastEventType, CastState, PlayerEventType } from 'react-native-theoplayer';
+import type { CastState } from 'react-native-theoplayer';
 import { PlayerContext, UiContext } from '../util/PlayerContext';
 import { DEFAULT_ACTION_BUTTON_STYLE } from './actionbutton/ActionButton';
+import { useChromecast } from '../../hooks/useChromecast';
+import type { ButtonBaseProps } from './ButtonBaseProps';
+import { TestIDs } from '../../utils/TestIDs';
 
-interface CastButtonState {
-  connected: boolean;
-}
-
-export interface CastButtonProps {
+export interface CastButtonProps extends ButtonBaseProps {
   /**
    * The icon component used in the button. Only overrideable for web.
    */
@@ -24,44 +22,18 @@ export function isConnected(state: CastState | undefined): boolean {
  * The native button to enable Chromecast for the `react-native-theoplayer` UI.
  * This component uses the button from `react-native-google-cast` and is not supported on web.
  */
-export class ChromecastButton extends PureComponent<CastButtonProps, CastButtonState> {
-  private static initialState: CastButtonState = {
-    connected: false,
-  };
+export function ChromecastButton(props: CastButtonProps) {
+  const castState = useChromecast();
 
-  constructor(props: CastButtonProps) {
-    super(props);
-    this.state = ChromecastButton.initialState;
-  }
-
-  componentDidMount() {
-    const player = (this.context as UiContext).player;
-    player.addEventListener(PlayerEventType.CAST_EVENT, this.onCastStateChangeEvent);
-    this.setState({ connected: player.cast.chromecast?.casting ?? false });
-  }
-
-  componentWillUnmount() {
-    const player = (this.context as UiContext).player;
-    player.removeEventListener(PlayerEventType.CAST_EVENT, this.onCastStateChangeEvent);
-  }
-
-  private onCastStateChangeEvent = (event: CastEvent) => {
-    if (event.subType != CastEventType.CHROMECAST_STATE_CHANGE) {
-      return;
-    }
-    this.setState({ connected: isConnected(event.state) });
-  };
-
-  render() {
-    const { connected } = this.state;
-    return (
-      <PlayerContext.Consumer>
-        {(context: UiContext) => (
-          <CastButton style={DEFAULT_ACTION_BUTTON_STYLE} tintColor={connected ? context.style.colors.iconSelected : context.style.colors.icon} />
-        )}
-      </PlayerContext.Consumer>
-    );
-  }
+  return (
+    <PlayerContext.Consumer>
+      {(context: UiContext) => (
+        <CastButton
+          style={props.style ?? DEFAULT_ACTION_BUTTON_STYLE}
+          testID={props.testID ?? TestIDs.CHROMECAST_BUTTON}
+          tintColor={isConnected(castState) ? context.style.colors.iconSelected : context.style.colors.icon}
+        />
+      )}
+    </PlayerContext.Consumer>
+  );
 }
-
-ChromecastButton.contextType = PlayerContext;