+++ src/ui/components/seekbar/SeekBar.tsx
@@ -2,21 +2,36 @@ import React, { useContext, useState } from 'react';
 import { type LayoutChangeEvent, StyleProp, View, ViewStyle } from 'react-native';
 import { PlayerContext, UiContext } from '../util/PlayerContext';
 import { Slider } from '@miblanchard/react-native-slider';
-import { useDuration } from '../hooks/useDuration';
-import { useSeekable } from '../hooks/useSeekable';
-import { useDebounce } from '../hooks/useDebounce';
+import { useDuration } from '../../hooks/useDuration';
+import { useSeekable } from '../../hooks/useSeekable';
+import { useDebounce } from '../../hooks/useDebounce';
 import { SingleThumbnailView } from './thumbnail/SingleThumbnailView';
 import { useSliderTime } from './useSliderTime';
+import { TestIDs } from '../../utils/TestIDs';
 
 export interface SeekBarProps {
   /**
    * Optional style applied to the SeekBar.
    */
   style?: StyleProp<ViewStyle>;
-
+  /**
+   * Optional style applied to the SeekBar container.
+   */
   sliderContainerStyle?: ViewStyle;
-
+  /**
+   * Optional style applied to the track left of the thumb.
+   */
+  sliderMinimumTrackStyle?: ViewStyle;
+  /**
+   * Optional style applied to the track right of the thumb.
+   */
   sliderMaximumTrackStyle?: ViewStyle;
+  /**
+   * An id used to locate this view in end-to-end tests.
+   *
+   * @default 'seek-bar'
+   */
+  testID?: string;
 }
 
 /**
@@ -65,6 +80,7 @@ export const SeekBar = (props: SeekBarProps) => {
       {(context: UiContext) => (
         <View
           style={[props.style ?? { flex: 1 }]}
+          testID={props.testID ?? TestIDs.SEEK_BAR}
           onLayout={(event: LayoutChangeEvent) => {
             setWidth(event.nativeEvent.layout.width);
           }}>
@@ -73,6 +89,7 @@ export const SeekBar = (props: SeekBarProps) => {
             minimumValue={seekableStart}
             maximumValue={seekableEnd}
             containerStyle={props.sliderContainerStyle ?? { marginHorizontal: 8 }}
+            minimumTrackStyle={props.sliderMinimumTrackStyle ?? {}}
             maximumTrackStyle={props.sliderMaximumTrackStyle ?? {}}
             step={1000}
             renderAboveThumbComponent={(_index: number, value: number) => {