+++ src/ui/utils/TestIDs.ts
@@ -1,0 +1,11 @@
+export const TestIDs = {
+  AIRPLAY_BUTTON: 'airplay-button',
+  CHROMECAST_BUTTON: 'chromecast-button',
+  FULLSCREEN_BUTTON: 'fullscreen-button',
+  MUTE_BUTTON: 'mute-button',
+  PIP_BUTTON: 'pip-button',
+  PLAY_BUTTON: 'play-button',
+  SEEK_BAR: 'seek-bar',
+  SKIP_FWD_BUTTON: 'skip-fwd-button',
+  SKIP_BWD_BUTTON: 'skip-bwd-button',
+};